delete directories switch

+++ projekt/test_patterns.py
@@ -7,10 +7,13 @@ import ollama_api
 import openai
 import functions
 
-if os.path.exists(functions.pattern_test_dir_output_path):
-    shutil.rmtree(functions.pattern_test_dir_output_path)
-if os.path.exists(functions.pattern_test_object_dir_output_path):
-    shutil.rmtree(functions.pattern_test_object_dir_output_path)
+delete_dirs = False
+
+if delete_dirs:
+    if os.path.exists(functions.pattern_test_dir_output_path):
+        shutil.rmtree(functions.pattern_test_dir_output_path)
+    if os.path.exists(functions.pattern_test_object_dir_output_path):
+        shutil.rmtree(functions.pattern_test_object_dir_output_path)
 
 pattern1G_OcrEn = "Get me the list of goods from picture. Show the address, date, time and name of company. When you find the phone number show it too. When you find the fax number show it too as fax_number. When you find you find the table number, the information about guest or order number show it too. Show me the output as JSON. The company name put in key company, the address of company in key address, phone number in key phone_number, server name in key server, station number in key station, order number in key order_number, table info in key table, number of guests in key guests, subtotal price to key sub_total, tax in key tax, total cost in key total, date in key date, time in key time. Every good name will be as key of the JSON in key goods and value of the good will be the another JSON with amount of goods in key amount and the cost of the good in key price."
 pattern1G_OcrCz = "Zjisti mi ze snímku seznam zboží. Vrať mi adresu, datum, čas a název společnosti. Když najdeš telefonní číslo, vrať ho také. Když najdeš faxové číslo, vrať ho také jako fax_number. Když najdeš číslo stolu, informace o počtu hostů nebo číslo objednávky, vrať je také. Výstup mi ukaž jako JSON. Název společnosti navrať pod klíčem company, adresu společnosti pod klíčem address, telefonní číslo pod klíčem phone_number, jméno číšníka pod klíčem server, číslo stanice pod klíčem station, číslo objednávky pod klíčem order_number, informace o stole pod klíčem table, počet hostů pod klíčem guests, mezisoučet ceny pod klíčem sub_total, daň pod klíčem tax, celkovou cenu pod klíčem total, datum pod klíčem date, čas pod klíčem time. Každý název zboží bude jako klíč JSON v klíči goods a hodnota zboží bude další JSON s množstvím zboží v klíči amount a cenou zboží v klíči price."
